[ major update ] No more statements; just expressions.
We have made function bodies expression based. This addresses issues of early returns and final expressions, but side stepping the issues.

Unit expressions can be sequenced, as we have expression _blocks_, however this resulted in a bit more of a bloated surface syntax.

I would like to resolve the surface syntax to be more neater, but this is extra non-important work.
Perhaps we can have variants on the surface language, and perhaps a shunting algorithm for operators not stolen from Idris (or at least extracted from idris itself).

Other changes include a really generic AST and a smarter way to deal with builtins.