[ context arguments ] assign information to roles.
when launching a session we need to be able to assign things to roles i.e. which processes they are....

our context handler requires this information internally so feed it in when launching a session.

this provides the intrinsic syntax to do so and ensure that role assignments are unique and only involve participants active in the role.