[ partial ] progress on runtime.
Provided runtime code for many of the terms.
The sending and receiving is left to do.
This will require marshalling of data, and restrictions that the type
is not a reference.

Nasty hack with recursion; we provide a runtime check that our stored
copy of the local stack has an address space that is a subset of the
one we currently have.
There needs to be a better way..