[ new feature ] Merging for synthesis.
First working iteration of Merging for Synthesis.

The decision procedure is sound but not complete.

Need to work out how to regain completeness.

Current implementation produces errors with merging which we, unfortunatly, did not have before.
We need to add better error messages to work out what is going on.
Hopefully, our soundess conditions _are_ correct but are not completely correct...