[ finally ] taken into account holes.
Holes complicate BiDirectional Typing of Sessions.

In our setting, holes force you to push typing information top-down in one direction so that when a hole is encountered you can report the expected type.
Perhaps having explicit mode changes woudl help.

An easy win that we shall do, is to change the elaborator to try and synthesise the type of a session, then check it.
SUch that if holes are there we fail and then top down, otherwise we synth.