[ features ] Better code synthesis and hole.
We can now display all holes in function and session bodies.

We are now capable of type-drivien term synthesis for typed sessions.
Thus supporting better interactive development.

Holes must be unique, need to add checks for that.

Made improvements to the UX of the REPL.

Deprecated `:run` in favour of `:exec`.